requirements.txt
anthropic
streamlit
pymupdf
python-docx
